Move SqlTransform external builder to it's own package, pull in data catalog table provider
