+++ sdks/java/extensions/sql/src/main/java/org/apache/beam/sdk/extensions/sql/SqlTransform.java
@@ -17,7 +17,6 @@
  */
 package org.apache.beam.sdk.extensions.sql;
 
-import com.google.auto.service.AutoService;
 import com.google.auto.value.AutoValue;
 import java.util.Collections;
 import java.util.HashMap;
@@ -25,7 +24,6 @@ import java.util.List;
 import java.util.Map;
 import javax.annotation.Nullable;
 import org.apache.beam.sdk.annotations.Experimental;
-import org.apache.beam.sdk.expansion.ExternalTransformRegistrar;
 import org.apache.beam.sdk.extensions.sql.impl.BeamSqlEnv;
 import org.apache.beam.sdk.extensions.sql.impl.BeamSqlEnv.BeamSqlEnvBuilder;
 import org.apache.beam.sdk.extensions.sql.impl.BeamSqlPipelineOptions;
@@ -36,7 +34,6 @@ import org.apache.beam.sdk.extensions.sql.meta.BeamSqlTable;
 import org.apache.beam.sdk.extensions.sql.meta.provider.ReadOnlyTableProvider;
 import org.apache.beam.sdk.extensions.sql.meta.provider.TableProvider;
 import org.apache.beam.sdk.transforms.Combine;
-import org.apache.beam.sdk.transforms.ExternalTransformBuilder;
 import org.apache.beam.sdk.transforms.PTransform;
 import org.apache.beam.sdk.transforms.SerializableFunction;
 import org.apache.beam.sdk.values.PCollection;
@@ -258,8 +255,8 @@ public abstract class SqlTransform extends PTransform<PInput, PCollection<Row>> 
   }
 
   @AutoValue.Builder
-  abstract static class Builder
-      implements ExternalTransformBuilder<External.Configuration, PInput, PCollection<Row>> {
+  abstract static class Builder {
+
     abstract Builder setQueryString(String queryString);
 
     abstract Builder setQueryParameters(QueryParameters queryParameters);
@@ -275,19 +272,6 @@ public abstract class SqlTransform extends PTransform<PInput, PCollection<Row>> 
     abstract Builder setDefaultTableProvider(@Nullable String defaultTableProvider);
 
     abstract SqlTransform build();
-
-    @Override
-    public PTransform<PInput, PCollection<Row>> buildExternal(
-        External.Configuration configuration) {
-      return builder()
-          .setQueryString(configuration.query)
-          .setQueryParameters(QueryParameters.ofNone())
-          .setUdafDefinitions(Collections.emptyList())
-          .setUdfDefinitions(Collections.emptyList())
-          .setTableProviderMap(Collections.emptyMap())
-          .setAutoUdfUdafLoad(false)
-          .build();
-    }
   }
 
   @AutoValue
@@ -311,26 +295,6 @@ public abstract class SqlTransform extends PTransform<PInput, PCollection<Row>> 
 
     static UdafDefinition of(String udafName, Combine.CombineFn combineFn) {
       return new AutoValue_SqlTransform_UdafDefinition(udafName, combineFn);
-    }
-  }
-
-  @AutoService(ExternalTransformRegistrar.class)
-  public static class External implements ExternalTransformRegistrar {
-
-    private static final String URN = "beam:external:java:sql:v1";
-
-    @Override
-    public Map<String, Class<? extends ExternalTransformBuilder>> knownBuilders() {
-      return org.apache.beam.vendor.guava.v26_0_jre.com.google.common.collect.ImmutableMap.of(
-          URN, AutoValue_SqlTransform.Builder.class);
-    }
-
-    public static class Configuration {
-      String query;
-
-      public void setQuery(String query) {
-        this.query = query;
-      }
     }
   }
 }
